feat(web): add full Flutter web platform support
Migrate 64-bit integer types from int to BigInt to address JavaScript's
53-bit precision limitation. This enables correct handling of all Stellar
and Soroban values on web platforms.

XDR layer:
- XdrInt64, XdrUint64: internal storage changed from int to BigInt
- XdrInt128Parts, XdrUInt128Parts: forHiLo() now takes BigInt parameters
- XdrInt256Parts, XdrUInt256Parts: forHiHiHiLoLoHiLoLo() now takes BigInt
- XdrSCVal: forU64, forI64, forTimepoint, forDuration now take BigInt
- XdrDataIO: web-safe readBigInt64/writeBigInt64 using byte manipulation
- Added readBigInt64Signed() for proper sign extension

High-level APIs:
- MemoId: constructor and id property now use BigInt
- MuxedAccount.id: changed to BigInt?
- AccountResponse.muxedAccountMed25519Id: changed to BigInt?
- SubmitTransactionResponse.getOfferIdFromResult(): returns BigInt?
- TimeBounds, FeeBumpTransaction, Claimant predicates: updated for BigInt
- Offer operations: use BigInt.parse() for offer IDs
- revokeOfferSponsorship(): offerId parameter changed to BigInt
- XdrLedgerKey.forOffer(), XdrLedgerKeyOffer.forOfferId(): offerId changed to BigInt
- XdrLedgerKey.getOfferOfferId(): returns BigInt?
- SorobanAddressCredentials.nonce: changed to BigInt
- SorobanCredentials.forAddress(): nonce parameter changed to BigInt

Platform abstractions:
- StellarSDK: new HTTP client abstraction via stub files (web/native)
- SorobanServer: platform-specific HTTP configuration via stubs
- Util.readFile(): throws UnsupportedError on web
- EventSourceEncoder: uses archive package, streaming gzip throws UnsupportedError

SEP updates:
- SEP-7 (URIScheme): BigInt memo handling
- SEP-10 (webauth): BigInt memo validation
- SEP-11 (txrep): BigInt.parse() for memo IDs and Soroban values

Soroban:
- ContractSpec: web-safe BigInt constants for i64 bounds

Tests:
- big_int_helpers_test.dart: comprehensive rewrite for BigInt verification
- All Soroban tests: use loadContractCode() for cross-platform WASM loading
- Updated all test files for BigInt comparisons

Dependencies:
- Added: archive ^3.6.1 (cross-platform gzip)
- Removed: fixnum package entirely (Int64 replaced with BigInt, Int32 with native int)
- Added: INT32_MAX_VALUE constant to BitConstants
- Added: platform-specific stub files for HTTP clients

BREAKING CHANGES: See v3_migration_guide.md for details.

+++ lib/src/muxed_account.dart
@@ -61,13 +61,13 @@ import 'dart:typed_data';
 class MuxedAccount {
   late String _accountId;
   String _ed25519AccountId;
-  int? _id;
+  BigInt? _id;
 
   /// Creates a MuxedAccount with an Ed25519 account ID and optional muxing ID.
   ///
   /// Parameters:
   /// - [_ed25519AccountId]: The underlying Ed25519 account ID (G... address)
-  /// - [_id]: Optional 64-bit multiplexing ID, or null for standard accounts
+  /// - [_id]: Optional 64-bit multiplexing ID as BigInt, or null for standard accounts
   MuxedAccount(this._ed25519AccountId, this._id) {
     _accountId = "0";
   }
@@ -129,10 +129,10 @@ class MuxedAccount {
   /// Returns: The base account ID without muxing information
   String get ed25519AccountId => _ed25519AccountId;
 
-  /// Gets the 64-bit multiplexing ID.
+  /// Gets the 64-bit multiplexing ID as BigInt.
   ///
   /// Returns: The ID if this is a muxed account, null for standard accounts
-  int? get id => _id;
+  BigInt? get id => _id;
 
   /// Gets the full account ID in the appropriate format.
   ///
@@ -190,7 +190,7 @@ class MuxedAccount {
   /// Returns: A [MuxedAccount] instance
   static MuxedAccount fromXdr(XdrMuxedAccount xdrMuxedAccount) {
     String? ed25519AccountId;
-    int? id;
+    BigInt? id;
     if (xdrMuxedAccount.discriminant ==
         XdrCryptoKeyType.KEY_TYPE_MUXED_ED25519) {
       ed25519AccountId = StrKey.encodeStellarAccountId(
